Add SequentialTransform
