+++ Project.toml
@@ -3,6 +3,9 @@ uuid = "756644eb-7399-463e-a86e-97605fb648cd"
 authors = ["Júlio Hoffimann <julio.hoffimann@gmail.com> and contributors"]
 version = "0.2.1"
 
+[deps]
+AbstractTrees = "1520ce14-60c1-5f80-bbc7-55ef81b5835c"
+
 [compat]
 julia = "1"
