+++ test/interface.jl
@@ -1,0 +1,2 @@
+@testset "Interface" begin
+end
