+++ test/runtests.jl
@@ -1,6 +1,14 @@
 using TransformsAPI
 using Test
 
+# list of tests
+testfiles = [
+  "interface.jl",
+  "sequential.jl"
+]
+
 @testset "TransformsAPI.jl" begin
-    # Write your tests here.
+  for testfile in testfiles
+    include(testfile)
+  end
 end
